Break up omdata.h, extract oms code into overmap_special.h
