+++ src/gamemode_defense.h
@@ -6,10 +6,10 @@
 #include <vector>
 
 #include "calendar.h"
+#include "coordinates.h"
 #include "cursesdef.h"
 #include "enums.h"
 #include "gamemode.h"
-#include "omdata.h"
 #include "point.h"
 #include "type_id.h"
