+++ src/omdata.h
@@ -18,6 +18,7 @@
 #include "numeric_interval.h"
 #include "coordinates.h"
 #include "int_id.h"
+#include "om_direction.h"
 #include "point.h"
 #include "string_id.h"
 #include "translations.h"
@@ -39,89 +40,6 @@ static const overmap_land_use_code_id land_use_code_forest( "forest" );
 static const overmap_land_use_code_id land_use_code_wetland( "wetland" );
 static const overmap_land_use_code_id land_use_code_wetland_forest( "wetland_forest" );
 static const overmap_land_use_code_id land_use_code_wetland_saltwater( "wetland_saltwater" );
-
-/** Direction on the overmap. */
-namespace om_direction
-{
-/** Basic enum for directions. */
-enum class type : int {
-    invalid = -1,
-    none,
-    north = none,
-    east,
-    south,
-    west,
-};
-
-/** For the purposes of iteration. */
-const std::array<type, 4> all = {{ type::north, type::east, type::south, type::west }};
-const size_t size = all.size();
-
-const std::array<std::string, 4> all_suffixes = {{ "_north", "_east", "_south", "_west" }};
-const std::string invalid_dir_suffix;
-const std::array<int, 4> all_cw_rotations = {{ 0, 1, 2, 3 }};
-const int invalid_dir_rotations = 0;
-
-/** Returns directional suffix associated with the value, e.g. _north or _west. */
-constexpr const std::string &get_suffix( type dir )
-{
-    if( dir == type::invalid ) {
-        return invalid_dir_suffix;
-    } else {
-        return all_suffixes[static_cast<size_t>( dir )];
-    }
-}
-
-/** Returns number of clockwise rotations needed to reach this direction from 'north'. */
-constexpr int get_num_cw_rotations( type dir )
-{
-    if( dir == type::invalid ) {
-        return invalid_dir_rotations;
-    } else {
-        return all_cw_rotations[static_cast<size_t>( dir )];
-    }
-}
-
-/** Number of bits needed to store directions. */
-const size_t bits = static_cast<size_t>( -1 ) >> ( CHAR_BIT *sizeof( size_t ) - size );
-
-/** Identifier for serialization purposes. */
-const std::string &id( type dir );
-
-/** Get Human readable name of a direction */
-std::string name( type dir );
-
-/** Various rotations. */
-point rotate( const point &p, type dir );
-tripoint rotate( const tripoint &p, type dir );
-uint32_t rotate_symbol( uint32_t sym, type dir );
-
-/** Returns point(0, 0) displaced in specified direction by a specified distance
- * @param dir Direction of displacement
- * @param dist Distance of displacement
- */
-point displace( type dir, int dist = 1 );
-
-/** Returns a sum of two numbers
- *  @param dir1 first number
- *  @param dir2 second number */
-type add( type dir1, type dir2 );
-
-/** Turn by 90 degrees to the left, to the right, or randomly (either left or right). */
-type turn_left( type dir );
-type turn_right( type dir );
-type turn_random( type dir );
-
-/** Returns an opposite direction. */
-type opposite( type dir );
-
-/** Returns a random direction. */
-type random();
-
-/** Whether these directions are parallel. */
-bool are_parallel( type dir1, type dir2 );
-
-} // namespace om_direction
 
 class overmap_land_use_code
 {
@@ -154,11 +72,7 @@ struct overmap_spawns {
         }
 
     protected:
-        template<typename JsonObjectType>
-        void load( JsonObjectType &jo ) {
-            jo.read( "group", group );
-            jo.read( "population", population );
-        }
+        void deserialize( const JsonObject &jo );
 };
 
 struct overmap_static_spawns : public overmap_spawns {
@@ -168,12 +82,7 @@ struct overmap_static_spawns : public overmap_spawns {
         return overmap_spawns::operator==( rhs ) && chance == rhs.chance;
     }
 
-    template<typename JsonStream>
-    void deserialize( JsonStream &jsin ) {
-        auto jo = jsin.get_object();
-        overmap_spawns::load( jo );
-        jo.read( "chance", chance );
-    }
+    void deserialize( const JsonObject &jo );
 };
 
 //terrain flags enum! this is for tracking the indices of each flag.
@@ -394,109 +303,6 @@ struct oter_t {
 bool operator==( const oter_id &lhs, const char *rhs );
 bool operator!=( const oter_id &lhs, const char *rhs );
 
-// LINE_**** corresponds to the ACS_**** macros in ncurses, and are patterned
-// the same way; LINE_NESW, where X indicates a line and O indicates no line
-// (thus, LINE_OXXX looks like 'T'). LINE_ is defined in output.h.  The ACS_
-// macros can't be used here, since ncurses hasn't been initialized yet.
-
-// Overmap specials--these are "special encounters," dungeons, nests, etc.
-// This specifies how often and where they may be placed.
-
-// OMSPEC_FREQ determines the length of the side of the square in which each
-// overmap special will be placed.  At OMSPEC_FREQ 6, the overmap is divided
-// into 900 squares; lots of space for interesting stuff!
-static constexpr int OMSPEC_FREQ = 15;
-
-struct overmap_special_spawns : public overmap_spawns {
-    numeric_interval<int> radius;
-
-    bool operator==( const overmap_special_spawns &rhs ) const {
-        return overmap_spawns::operator==( rhs ) && radius == rhs.radius;
-    }
-
-    template<typename JsonStream>
-    void deserialize( JsonStream &jsin ) {
-        auto jo = jsin.get_object();
-        overmap_spawns::load( jo );
-        jo.read( "radius", radius );
-    }
-};
-
-struct overmap_special_terrain {
-    overmap_special_terrain() = default;
-    tripoint p;
-    oter_str_id terrain;
-    std::set<std::string> flags;
-    std::set<overmap_location_id> locations;
-
-    template<typename JsonStream>
-    void deserialize( JsonStream &jsin ) {
-        auto om = jsin.get_object();
-        om.read( "point", p );
-        om.read( "overmap", terrain );
-        om.read( "flags", flags );
-        om.read( "locations", locations );
-    }
-
-    /**
-     * Returns whether this terrain of the special can be placed on the specified terrain.
-     * It's true if oter meets any of locations.
-     */
-    bool can_be_placed_on( const oter_id &oter ) const;
-};
-
-struct overmap_special_connection {
-    tripoint p;
-    cata::optional<tripoint> from;
-    om_direction::type initial_dir = om_direction::type::invalid;
-    // TODO: Remove it.
-    oter_type_str_id terrain;
-    overmap_connection_id connection;
-    bool existing = false;
-
-    template<typename JsonStream>
-    void deserialize( JsonStream &jsin ) {
-        auto jo = jsin.get_object();
-        jo.read( "point", p );
-        jo.read( "terrain", terrain );
-        jo.read( "existing", existing );
-        jo.read( "connection", connection );
-        assign( jo, "from", from );
-    }
-};
-
-class overmap_special
-{
-    public:
-        /** Returns terrain at the given point. */
-        const overmap_special_terrain &get_terrain_at( const tripoint &p ) const;
-        /** @returns true if this special requires a city */
-        bool requires_city() const;
-        /** @returns whether the special at specified tripoint can belong to the specified city. */
-        bool can_belong_to_city( const tripoint_om_omt &p, const city &cit ) const;
-
-        overmap_special_id id;
-        std::list<overmap_special_terrain> terrains;
-        std::vector<overmap_special_connection> connections;
-
-        numeric_interval<int> city_size{ 0, INT_MAX };
-        numeric_interval<int> city_distance{ 0, INT_MAX };
-        numeric_interval<int> occurrences;
-
-        bool rotatable = true;
-        overmap_special_spawns spawns;
-        std::set<std::string> flags;
-
-        // Used by generic_factory
-        bool was_loaded = false;
-        void load( const JsonObject &jo, const std::string &src );
-        void finalize();
-        void check() const;
-    private:
-        // These locations are the default values if ones are not specified for the individual OMTs.
-        std::set<overmap_location_id> default_locations;
-};
-
 namespace overmap_terrains
 {
 
@@ -520,30 +326,5 @@ void reset();
 const std::vector<overmap_land_use_code> &get_all();
 
 } // namespace overmap_land_use_codes
-
-namespace overmap_specials
-{
-
-void load( const JsonObject &jo, const std::string &src );
-void finalize();
-void check_consistency();
-void reset();
-
-const std::vector<overmap_special> &get_all();
-
-overmap_special_batch get_default_batch( const point_abs_om &origin );
-/**
- * Generates a simple special from a building id.
- */
-overmap_special_id create_building_from( const oter_type_str_id &base );
-
-} // namespace overmap_specials
-
-namespace city_buildings
-{
-
-void load( const JsonObject &jo, const std::string &src );
-
-} // namespace city_buildings
 
 #endif // CATA_SRC_OMDATA_H
