+++ src/overmap_ui.cpp
@@ -55,6 +55,7 @@
 #include "overmap.h"
 #include "overmap_types.h"
 #include "overmapbuffer.h"
+#include "overmap_special.h"
 #include "regional_settings.h"
 #include "rng.h"
 #include "sdltiles.h"
