+++ tests/test_main.cpp
@@ -45,6 +45,7 @@
 #include "options.h"
 #include "output.h"
 #include "overmap.h"
+#include "overmap_special.h"
 #include "overmapbuffer.h"
 #include "path_info.h"
 #include "pldata.h"
